Changed Post to PostForm.jsx and added LoadPost for later
