+++ Binturong/src/pages/UserFeed.jsx
@@ -1,10 +1,14 @@
 //Start of the User Feed page
 
-import PostForm from "../Components/Post"
+import PostForm from "../Components/PostForm"
 
 const UserFeed = () => {
 
- return(<PostForm />);
+
+ return(
+ <PostForm />
+
+); 
 }
 
   export default UserFeed;
