+++ binturong-api/src/main/java/model/Post.java
@@ -1,0 +1,14 @@
+package model;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+
+@Entity
+public class Post {
+    @Id
+    @GeneratedValue
+    private int id;
+
+
+}
